third_party/blink/web_tests/http/tests/devtools/console/console-search-reveals-messages-expected.txt

Tests that console viewport reveals messages on searching.


Running: waitForMessages
Message count: 201

Running: verifyViewportIsTallEnough

Running: scrollConsoleToTop
first visible message index: 0

Running: testFindLastMessage
last visible message index: 200
